pyproject.toml
[project]
name = "hn-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm (>=1.59.6,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "transformers (>=4.48.1,<5.0.0)",
    "torch (>=2.5.1,<3.0.0)",
    "fire (>=0.7.0,<0.8.0)",
    "openai (>=1.60.1,<2.0.0)",
    "optuna (>=4.2.0,<5.0.0)",
    "selenium (>=4.28.1,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

